Feedback from review: "Should imo be moved to a types.ts file in the directory"
